fix: avoid to fail if crd not exists (#30)

+++ scripts/remove-dependencies-pre.sh
@@ -2,7 +2,11 @@
 oc delete --ignore-not-found kedacontroller keda -n openshift-keda
 
 # Remove NFD CR
-oc delete --ignore-not-found nodefeaturediscovery nfd-instance -n openshift-nfd
+if oc get crd nodefeaturediscoveries.nfd.openshift.io &>/dev/null; then
+  oc delete --ignore-not-found nodefeaturediscovery nfd-instance -n openshift-nfd
+fi
 
 # Remove GPU operator ClusterPolicy CR
-oc delete --ignore-not-found clusterpolicy gpu-cluster-policy
+if oc get crd clusterpolicies.nvidia.com &>/dev/null; then
+  oc delete --ignore-not-found clusterpolicy gpu-cluster-policy
+fi
